requirements/base.txt
django==2.1.15
djangorestframework==3.8.0
markdown==3.1.1
django-filter==2.2.0
coreapi==2.3.3  
Pygments==2.5.2
